source/requirements.txt
altgraph==0.17
appdirs==1.4.4
Bottleneck==1.3.2
certifi==2020.6.20
cycler==0.10.0
decorator==4.4.2
fbs==0.9.0
ffmpeg-python==0.2.0
future==0.18.2
imageio==2.9.0
kiwisolver==1.2.0
macholib==1.14
matplotlib==3.2.2
networkx==2.5
numpy==1.19.3
pandas==1.1.4
pefile==2019.4.18
Pillow==8.0.1
PIMS==0.5
PyInstaller==3.4
pyparsing==2.4.7
PyQt5==5.9.2
pyqtgraph==0.11.0
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.17.2
scipy==1.4.1
sip==4.19.8
six==1.15.0
slicerator==1.0.0
tifffile==2020.10.1
trackpy==0.4.2
